pyproject.toml
[project]
name = "multi-agent-system"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "agno>=1.7.7",
    "db-sqlite3>=0.0.1",
    "fastapi>=0.116.1",
    "gcloud>=0.18.3",
    "google-genai>=1.28.0",
    "openai>=1.98.0",
    "python-dotenv>=1.1.1",
    "ag-ui-protocol==0.1.8",
    "uvicorn>=0.24.0",
]
